requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
Automat==25.4.16
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
cssselect==1.4.0
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.129.0
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
fastembed==0.7.4
filelock==3.21.0
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itemadapter==0.13.1
itemloaders==1.4.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.24.1
packaging==26.0
parsel==1.11.0
parso==0.8.6
pillow==11.3.0
prompt_toolkit==3.0.52
Protego==0.6.0
protobuf==6.33.5
psycopg2==2.9.11
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.0
pydantic_core==2.41.5
PyDispatcher==2.0.7
Pygments==2.19.2
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
queuelib==1.9.0
requests==2.32.5
requests-file==3.0.1
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.7.6
Scrapy==2.14.1
sentry-sdk==2.53.0
service-identity==24.2.0
shellingham==1.5.4
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tldextract==5.3.1
tokenizers==0.22.2
tqdm==4.67.3
traitlets==5.14.3
Twisted==25.5.0
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
w3lib==2.4.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
win32_setctime==1.2.0
zope.interface==8.2
